requirements.txt
numpy<2.0.0
torch>=1.13.1,<2.0.0
transformers==4.30.2
pybind11>=2.12.0
safetensors>=0.3.1
matplotlib>=3.4.0
pandas>=1.3.0
pyyaml>=5.4.0
typing-extensions>=4.0.0
tqdm>=4.62.0
scikit-learn>=0.24.0
